Refactored code generator
